Renamed directory name of actions/joinRoom to actions/helpers
