+++ src/actions/joinRoom.js
@@ -1,11 +1,9 @@
-import socket from '../socketHandlers/index';
-import { joinRoomPromise } from './joinRoom/joinRoomHelper';
+import { joinRoomPromise } from './helpers/joinRoomHelper';
 
 const SOCKET_EVENT_TYPES = {
   EMIT: 'joinRoom',
   ON: 'resultJoinRoom'
 };
-
 
 const joinRoom = (roomId) => {
   joinRoomPromise(roomId);
@@ -19,17 +17,5 @@ const joinRoomSuccess = (roomId) => ({
   type: 'JOIN_ROOM',
   roomId
 });
-
-// The reason I created this function is for Test.
-// Mocking this function, I can write a test for joinRoom.
-// const joinRoomPromise = (roomId) => {
-//   return new Promise((resolve, reject) => {
-//     socket.on(SOCKET_EVENT_TYPES.ON, (data) => {
-//       resolve(data);
-//     });
-//     socket.emit(SOCKET_EVENT_TYPES.EMIT, {roomId});
-//   });
-// };
-
 
 export { joinRoom };
